refactor: rearranged ground svc page
