+++ fbw-a32nx/src/systems/instruments/src/EFB/Store/features/groundServicePage.ts
@@ -1,4 +1,5 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
+import { airframeType } from '../../Efb';
 
 enum ServiceButtonState {
     HIDDEN,
@@ -10,45 +11,60 @@ enum ServiceButtonState {
 }
 
 interface ButtonSelectionState {
-    cabinLeftDoorButtonState: ServiceButtonState
-    cabinRightDoorButtonState: ServiceButtonState
+    boarding1DoorButtonState: ServiceButtonState
+    boarding2DoorButtonState: ServiceButtonState
+    boarding3DoorButtonState: ServiceButtonState
+    serviceDoorButtonState: ServiceButtonState
+    cargo1DoorButtonState: ServiceButtonState
     jetWayButtonState: ServiceButtonState
     fuelTruckButtonState: ServiceButtonState
     gpuButtonState: ServiceButtonState
-    cargoDoorButtonState: ServiceButtonState
     baggageButtonState: ServiceButtonState
-    aftLeftDoorButtonState: ServiceButtonState
-    aftRightDoorButtonState: ServiceButtonState
     cateringButtonState: ServiceButtonState
 }
 
 // hack to fix initialization issue for ACE/vite
 let initialState: ButtonSelectionState = {
-    cabinLeftDoorButtonState: ServiceButtonState.DISABLED,
-    cabinRightDoorButtonState: ServiceButtonState.DISABLED,
-    jetWayButtonState: ServiceButtonState.DISABLED,
+    boarding1DoorButtonState: ServiceButtonState.DISABLED,
+    boarding2DoorButtonState: ServiceButtonState.DISABLED,
+    boarding3DoorButtonState: ServiceButtonState.DISABLED,
+    serviceDoorButtonState: ServiceButtonState.DISABLED,
+    cargo1DoorButtonState: ServiceButtonState.DISABLED,
     fuelTruckButtonState: ServiceButtonState.DISABLED,
     gpuButtonState: ServiceButtonState.DISABLED,
-    cargoDoorButtonState: ServiceButtonState.DISABLED,
+    jetWayButtonState: ServiceButtonState.DISABLED,
     baggageButtonState: ServiceButtonState.DISABLED,
-    aftLeftDoorButtonState: ServiceButtonState.DISABLED,
-    aftRightDoorButtonState: ServiceButtonState.DISABLED,
     cateringButtonState: ServiceButtonState.DISABLED,
 };
 
 const setInitialState = () => {
-    initialState = {
-        cabinLeftDoorButtonState: (SimVar.GetSimVarValue('A:INTERACTIVE POINT OPEN:0', 'Percent over 100') === 1.0 ? ServiceButtonState.ACTIVE : ServiceButtonState.INACTIVE),
-        cabinRightDoorButtonState: (SimVar.GetSimVarValue('A:INTERACTIVE POINT OPEN:1', 'Percent over 100') === 1.0 ? ServiceButtonState.ACTIVE : ServiceButtonState.INACTIVE),
-        jetWayButtonState: ServiceButtonState.INACTIVE,
-        fuelTruckButtonState: (SimVar.GetSimVarValue('A:INTERACTIVE POINT OPEN:9', 'Percent over 100') === 1.0 ? ServiceButtonState.ACTIVE : ServiceButtonState.INACTIVE),
-        gpuButtonState: (SimVar.GetSimVarValue('A:INTERACTIVE POINT OPEN:8', 'Percent over 100') === 1.0 ? ServiceButtonState.ACTIVE : ServiceButtonState.INACTIVE),
-        cargoDoorButtonState: (SimVar.GetSimVarValue('A:INTERACTIVE POINT OPEN:5', 'Percent over 100') === 1.0 ? ServiceButtonState.ACTIVE : ServiceButtonState.INACTIVE),
-        baggageButtonState: ServiceButtonState.INACTIVE,
-        aftLeftDoorButtonState: (SimVar.GetSimVarValue('A:INTERACTIVE POINT OPEN:2', 'Percent over 100') === 1.0 ? ServiceButtonState.ACTIVE : ServiceButtonState.INACTIVE),
-        aftRightDoorButtonState: (SimVar.GetSimVarValue('A:INTERACTIVE POINT OPEN:3', 'Percent over 100') === 1.0 ? ServiceButtonState.ACTIVE : ServiceButtonState.INACTIVE),
-        cateringButtonState: ServiceButtonState.INACTIVE,
-    };
+    if (airframeType === 'A380-842') {
+        initialState = {
+            boarding1DoorButtonState: (SimVar.GetSimVarValue('A:INTERACTIVE POINT OPEN:0', 'Percent over 100') === 1.0 ? ServiceButtonState.ACTIVE : ServiceButtonState.INACTIVE),
+            boarding2DoorButtonState: (SimVar.GetSimVarValue('A:INTERACTIVE POINT OPEN:2', 'Percent over 100') === 1.0 ? ServiceButtonState.ACTIVE : ServiceButtonState.INACTIVE),
+            boarding3DoorButtonState: (SimVar.GetSimVarValue('A:INTERACTIVE POINT OPEN:10', 'Percent over 100') === 1.0 ? ServiceButtonState.ACTIVE : ServiceButtonState.INACTIVE),
+            serviceDoorButtonState: (SimVar.GetSimVarValue('A:INTERACTIVE POINT OPEN:9', 'Percent over 100') === 1.0 ? ServiceButtonState.ACTIVE : ServiceButtonState.INACTIVE),
+            cargo1DoorButtonState: (SimVar.GetSimVarValue('A:INTERACTIVE POINT OPEN:16', 'Percent over 100') === 1.0 ? ServiceButtonState.ACTIVE : ServiceButtonState.INACTIVE),
+            fuelTruckButtonState: (SimVar.GetSimVarValue('A:INTERACTIVE POINT OPEN:18', 'Percent over 100') === 1.0 ? ServiceButtonState.ACTIVE : ServiceButtonState.INACTIVE),
+            gpuButtonState: (SimVar.GetSimVarValue('A:INTERACTIVE POINT OPEN:19', 'Percent over 100') === 1.0 ? ServiceButtonState.ACTIVE : ServiceButtonState.INACTIVE),
+            jetWayButtonState: ServiceButtonState.INACTIVE,
+            baggageButtonState: ServiceButtonState.INACTIVE,
+            cateringButtonState: ServiceButtonState.INACTIVE,
+        };
+    } else {
+        initialState = {
+            boarding1DoorButtonState: (SimVar.GetSimVarValue('A:INTERACTIVE POINT OPEN:0', 'Percent over 100') === 1.0 ? ServiceButtonState.ACTIVE : ServiceButtonState.INACTIVE),
+            boarding2DoorButtonState: (SimVar.GetSimVarValue('A:INTERACTIVE POINT OPEN:1', 'Percent over 100') === 1.0 ? ServiceButtonState.ACTIVE : ServiceButtonState.INACTIVE),
+            boarding3DoorButtonState: (SimVar.GetSimVarValue('A:INTERACTIVE POINT OPEN:2', 'Percent over 100') === 1.0 ? ServiceButtonState.ACTIVE : ServiceButtonState.INACTIVE),
+            cargo1DoorButtonState: (SimVar.GetSimVarValue('A:INTERACTIVE POINT OPEN:5', 'Percent over 100') === 1.0 ? ServiceButtonState.ACTIVE : ServiceButtonState.INACTIVE),
+            serviceDoorButtonState: (SimVar.GetSimVarValue('A:INTERACTIVE POINT OPEN:3', 'Percent over 100') === 1.0 ? ServiceButtonState.ACTIVE : ServiceButtonState.INACTIVE),
+            fuelTruckButtonState: (SimVar.GetSimVarValue('A:INTERACTIVE POINT OPEN:9', 'Percent over 100') === 1.0 ? ServiceButtonState.ACTIVE : ServiceButtonState.INACTIVE),
+            gpuButtonState: (SimVar.GetSimVarValue('A:INTERACTIVE POINT OPEN:8', 'Percent over 100') === 1.0 ? ServiceButtonState.ACTIVE : ServiceButtonState.INACTIVE),
+            jetWayButtonState: ServiceButtonState.INACTIVE,
+            baggageButtonState: ServiceButtonState.INACTIVE,
+            cateringButtonState: ServiceButtonState.INACTIVE,
+        };
+    }
 };
 
 // hack to fix initialization issue for ACE/vite
@@ -62,11 +78,20 @@ export const buttonsSlice = createSlice({
     name: 'ground',
     initialState,
     reducers: {
-        setCabinLeftDoorButtonState: (state, action: PayloadAction<ServiceButtonState>) => {
-            state.cabinLeftDoorButtonState = action.payload;
+        setBoarding1DoorButtonState: (state, action: PayloadAction<ServiceButtonState>) => {
+            state.boarding1DoorButtonState = action.payload;
         },
-        setCabinRightDoorButtonState: (state, action: PayloadAction<ServiceButtonState>) => {
-            state.cabinRightDoorButtonState = action.payload;
+        setBoarding2DoorButtonState: (state, action: PayloadAction<ServiceButtonState>) => {
+            state.boarding2DoorButtonState = action.payload;
+        },
+        setBoarding3DoorButtonState: (state, action: PayloadAction<ServiceButtonState>) => {
+            state.boarding3DoorButtonState = action.payload;
+        },
+        setServiceDoorButtonState: (state, action: PayloadAction<ServiceButtonState>) => {
+            state.serviceDoorButtonState = action.payload;
+        },
+        setCargo1DoorButtonState: (state, action: PayloadAction<ServiceButtonState>) => {
+            state.cargo1DoorButtonState = action.payload;
         },
         setJetWayButtonState: (state, action: PayloadAction<ServiceButtonState>) => {
             state.jetWayButtonState = action.payload;
@@ -77,17 +102,8 @@ export const buttonsSlice = createSlice({
         setGpuButtonState: (state, action: PayloadAction<ServiceButtonState>) => {
             state.gpuButtonState = action.payload;
         },
-        setCargoDoorButtonState: (state, action: PayloadAction<ServiceButtonState>) => {
-            state.cargoDoorButtonState = action.payload;
-        },
         setBaggageButtonState: (state, action: PayloadAction<ServiceButtonState>) => {
             state.baggageButtonState = action.payload;
-        },
-        setAftLeftDoorButtonState: (state, action: PayloadAction<ServiceButtonState>) => {
-            state.aftLeftDoorButtonState = action.payload;
-        },
-        setAftRightDoorButtonState: (state, action: PayloadAction<ServiceButtonState>) => {
-            state.aftRightDoorButtonState = action.payload;
         },
         setCateringButtonState: (state, action: PayloadAction<ServiceButtonState>) => {
             state.cateringButtonState = action.payload;
@@ -96,15 +112,15 @@ export const buttonsSlice = createSlice({
 });
 
 export const {
-    setCabinLeftDoorButtonState,
-    setCabinRightDoorButtonState,
+    setBoarding1DoorButtonState,
+    setBoarding2DoorButtonState,
+    setBoarding3DoorButtonState,
+    setServiceDoorButtonState,
+    setCargo1DoorButtonState,
     setJetWayButtonState,
     setFuelTruckButtonState,
     setGpuButtonState,
-    setCargoDoorButtonState,
     setBaggageButtonState,
-    setAftLeftDoorButtonState,
-    setAftRightDoorButtonState,
     setCateringButtonState,
 } = buttonsSlice.actions;
